Uncomment camera_gmsl_publisher
Uncommented camera_gmsl_publisher due to invalid XML syntax.

It turns out that with the arguments for image rescaling we
can't comment out the camera part, as it leads to invalid XML
errors on startup. So far the solution is that if we don't want
the camera node to run, we just have to delete that line
from the file instead of commenting it.