PX2 performance testing branch merge (#6)
* Update real world launchfile

Updated real world launchfile with parameters
and correct launch of lidar node

* change Lidar serial number

Changed the serial number of the lidar to match the new
Ouster OS0-64 sensor.

* Add new launchfile for PX2 performance testing

* Change features in manager.py

* Update performance testing setup

* Disable faulty fix2tfpose

Disabled fix2tfpose as it was publishing NaN transforms
which caused TF errors

* Update code for PX2 performance tests

Various updates for testing PX2

* Disable some nodes + fix syntax error

* Change car speed feedback topic

Changed the topic for feeding back the car's speed,
now it uses the speed message published from the CAN bus.

* add error message publisher

* Add Process Killer Scripts (#2)

Added a script to kill remaining processes
after stopping the ROS with CTRL+C

* Add Process Killer Scripts (#2)

Added a script to kill remaining processes
after stopping the ROS with CTRL+C

* Publish goal only once

Changed goal_publisher.py so that the goal message gets
published only once, and not continuously.

* Update features for testing

* Performance test changes

* Add initial pose for current 'standard' starting position

* add parameter to change resolution

* Change IMU topic name to the correct one

Changed the IMU's topic name to match
the one used by the current driver.

* Uncomment camera_gmsl_publisher

Uncommented camera_gmsl_publisher due to invalid XML syntax.

It turns out that with the arguments for image rescaling we
can't comment out the camera part, as it leads to invalid XML
errors on startup. So far the solution is that if we don't want
the camera node to run, we just have to delete that line
from the file instead of commenting it.

* Remove lidar_timestamp_update.py from IRL launchfiles

Due to high load on CPU and no effective use.

For rosbag playbacks, I created the
manager_bag_performance_testing.launch
file, which still has the lidar timestamp updater until we
create a script for rewriting rosbags.

* Updated the initial coordinate next to Brizalia

* Modify voxel leaf size for downsampling lidar point clouds

* More precise initial pose for Brinellvagen.

* Add scripts to automate test runs and analyse log files, for localizer.

* Read origin GPS coordinates from map folder

Made changes so that GPS coordinates of the origin of the map
are not hardcoded anymore in an AD-EYE launchfile,
but are in a separate launchfile shipped with the map instead.

* Add variable to map file paths (#4)

Instead of hard-coded names

* Update localization parameters

Updated settings in my_localization.launch
and gnss_broadcaster.py

* Change Rviz goal publishing topic

Changed the topic for publishing the goal from Rviz
so that it publishes directly to move_base by
circumventing goal_sequencer

(may be reverted later)

* Change manager.py configuration for tests

* Change launchfiles for testing setup

---------

Co-authored-by: iozgen <iozgen@kth.se>
Co-authored-by: alalr <alalr@kth.se>
Co-authored-by: mengliu-kth <mengliu@kth.se>
Co-authored-by: mktomaszuk <tomaszuk@kth.se>