Add functionality to transform GNSS-based position to PCL map
Added the possibility to create a transform between the GNSS-based
position and the frame of the point cloud map.
If such a transform exists, gnss_broadcaster.py will convert
the GNSS-based position to the frame of the PCL map, so that the
GNSS measurements are aligned with the point cloud.

As of now, this transformation is defined per map, in the
map_params.launch file for each map.
In case no transformation is defined for a given map,
gnss_broadcaster.py will just default to the previous solution,
assuming that GNSS positions are already in map frame.