Update real world launchfile
Add the lidar to the real world launchfile,
and remove the camera to match the actual configuration
we use with the PX2.

Removed the _phy from vehicle_controller as we only use one
controller, so the distinction is not needed anymore.

For the same reason, removed _phy from current_velocity,
so the remapping of the velocity feedback topics is not needed
anymore, and the main velocity topic is /current_velocity.